// types/umkm.ts
interface Umkm {
  id: string;
  nama_pemilik: string;
  nomor_pemilik: string;
  alamat_pemilik: string;
  nama_usaha: string;
  alamat_usaha: string;
  domisili_usaha: string;
  kodePos_usaha: string;
  email_usaha: string;
  tahunBerdiri_usaha: string;
  jenisbadan_usaha: string;
  kategori_usaha: string;
  deskripsi_usaha: string;
  legalitas_usaha: string;
  image: string; 
}

export { type Umkm };